basic movement juiced

+++ src/consts.rs
@@ -20,7 +20,7 @@ pub const ZIX_SKELLY: f32 = 19.0;
 pub const ZIX_MAX: f32 = 600.0; // Anything further forward than this gets culled by the camera(s)
 pub const ZIX_MENU: i32 = 50;
 pub const ZIX_MIN: f32 = -600.0; // Anything further back than this gets culled by the camera(s)
-pub const ZIX_PARTICLES: i32 = 40;
+pub const ZIX_PARTICLES: f32 = 40.0;
 pub const ZIX_TRANSITION: i32 = 60;
 
 // Colors are easy!
